Add scan function, remove for loop in find_path()
Added a scan function, which handles manual, random, and sensor input.
Removed a for loop in find_path, which is suspected to be redundant, but
somehow behaved correctly.
This code, which is also in master, is outdated (not used in the final project),
as it contains no calls to Motor.py and shows teleporting behavior when
encountering a dead end.